New version of Emacs, compiling from Git. Also fixed mmm-mode and turned on the `svn:executable` property for a bunch of files that needed it.

+++ bin/emacs.sh
@@ -4,18 +4,20 @@ export VERSION=22.3
 
 . `dirname $0`/functions.sh
 
-setup /Applications/Emacs.app/Contents/MacOS/Emacs
-download http://ftp.gnu.org/pub/gnu/emacs/emacs-$VERSION.tar.gz
-rm -rf emacs-$VERSION
-tar zxf emacs-$VERSION.tar.gz
-cd emacs-$VERSION
-patch -p1 < `dirname $0`/../patches/emacs-22.1-10.5.patch
-#./configure --with-carbon --without-x
-./configure --with-carbon --without-x --without-pop --with-xpm --with-jpeg --with-tiff --with-png --with-gif --with-x-toolkit=lucid
-# Might need to apply http://article.gmane.org/gmane.emacs.bugs/16867
-make
-sudo make install
-cp -rf Mac/Emacs.app /Applications
+#setup /Applications/Emacs.app/Contents/MacOS/Emacs
+cd /usr/local/src
+if [ ! -d emacs ]; then
+    git clone git://git.sv.gnu.org/emacs || exit $?
+fi
+cd emacs
+git pull || exit $?
+./configure --with-carbon --without-x --without-pop --with-xpm --with-jpeg --with-tiff --with-png --with-gif --with-x-toolkit=lucid --with-ns || exit $?
+make bootstrap || exit $?
+make || exit $?
+sudo make install || exit $?
+cd nextstep
+make install || exit $?
+cp -rf Emacs.app /Applications
 rm /usr/local/bin/emacs
 echo '#!/bin/sh' > /usr/local/bin/emacs
 echo '/Applications/Emacs.app/Contents/MacOS/Emacs "$@"' >> /usr/local/bin/emacs
